# docs/api/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "Your Auth API",
    "description": "API de autenticacao e identidade para contas proprietarias, projetos, usuarios finais, sessoes, refresh tokens e API keys de projeto.",
    "version": "0.0.1-SNAPSHOT"
  },
  "servers": [
    {
      "url": "http://localhost:8080",
      "description": "Ambiente local"
    }
  ],
...
  "paths": {
    "/projects/{projectId}/users/{userId}": {
      "get": {
        "tags": [
          "Project Users"
        ],
        "summary": "Busca um usuario final pelo id",
        "operationId": "findById",
        "parameters": [
          {
            "name": "projectId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "userId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Usuario encontrado.",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/UserResponseDTO"
                }
              }
            }
          },
          "401": {
            "description": "Autenticacao obrigatoria.",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "403": {
            "description": "Conta sem acesso ao projeto.",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "404": {
            "description": "Projeto ou usuario nao encontrados.",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        },
        "security": [
          {
            "bearerAuth": []
          },
          {
            "accessTokenCookie": []
          }
        ]
      },
      "put": {
        "tags": [
          "Project Users"
        ],
        "summary": "Atualiza um usuario final",
        "operationId": "update",
        "parameters": [
          {
            "name": "projectId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "userId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateUserDTO"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Usuario atualizado.",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/UserResponseDTO"
                }
              }
            }
          },
          "400": {
            "description": "Corpo da requisicao invalido ou erro de validacao.",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "401": {
            "description": "Autenticacao obrigatoria.",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "403": {
            "description": "Conta sem acesso ao projeto.",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "404": {
            "description": "Projeto ou usuario nao encontrados.",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        },
        "security": [
          {
            "bearerAuth": []
          },
          {
            "accessTokenCookie": []
          }
        ]
      },
      "delete": {
        "tags": [
          "Project Users"
        ],
        "summary": "Remove um usuario final",
        "operationId": "delete",
        "parameters": [
          {
            "name": "projectId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "userId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "204": {
            "description": "Usuario removido."
          },
          "401": {
            "description": "Autenticacao obrigatoria.",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "403": {
            "description": "Conta sem acesso ao projeto.",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "404": {
            "description": "Projeto ou usuario nao encontrados.",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        },
        "security": [
          {
            "bearerAuth": []
          },
          {
            "accessTokenCookie": []
          }
        ]
      }
    },
...
    "/users": {
      "post": {
        "tags": [
          "Public Project Users"
        ],
        "summary": "Cria um usuario final via API key",
        "description": "Cria um usuario final no projeto associado a API key enviada.",
        "operationId": "create",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateUserDTO"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "description": "Usuario criado.",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/UserResponseDTO"
                }
              }
            }
          },
          "400": {
            "description": "Corpo da requisicao invalido, erro de validacao ou API key mal formatada.",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "401": {
            "description": "API key ausente, invalida, expirada ou revogada.",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "403": {
            "description": "API key sem permissao para criar usuarios.",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "409": {
            "description": "Usuario ja existente no projeto.",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "500": {
            "description": "Falha ao processar API key do projeto.",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        },
        "security": [
          {
            "projectApiKey": []
          }
        ]
      }
    },
    "/users/refresh": {
      "post": {
        "tags": [
          "Public Project Users"
        ],
        "summary": "Renova a sessao web de um usuario final",
        "description": "Usa o cookie refresh_token para gerar novos tokens e redefinir cookies HTTP-only.",
        "operationId": "refresh",
        "parameters": [
          {
            "name": "refresh_token",
            "in": "cookie",
            "description": "Refresh token HTTP-only recebido no login do usuario final.",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Sessao renovada. Novos cookies access-token e refresh_token sao enviados no header Set-Cookie.",
            "headers": {
              "Set-Cookie": {
                "description": "Define novos cookies HTTP-only access-token e refresh_token.",
                "style": "simple",
                "schema": {
                  "type": "string"
                }
              }
            }
          },
          "400": {
            "description": "Cookie refresh_token ausente.",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "401": {
            "description": "Refresh token invalido, expirado ou reutilizado.",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "404": {
            "description": "Usuario ou sessao associados ao refresh token nao encontrados.",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "500": {
            "description": "Falha ao validar, armazenar ou gerar tokens.",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        },
        "security": [
          {
            "refreshTokenCookie": []
          }
        ]
      }
    },
    "/users/logout": {
      "post": {
        "tags": [
          "Public Project Users"
        ],
        "summary": "Encerra a sessao web de um usuario final",
        "description": "Revoga a sessao associada ao refresh token e limpa os cookies da sessao.",
        "operationId": "logout",
        "parameters": [
          {
            "name": "refresh_token",
            "in": "cookie",
            "description": "Refresh token HTTP-only recebido no login do usuario final.",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "204": {
            "description": "Sessao encerrada e cookies limpos.",
            "headers": {
              "Set-Cookie": {
                "description": "Limpa os cookies HTTP-only access-token e refresh_token.",
                "style": "simple",
                "schema": {
                  "type": "string"
                }
              }
            }
          },
          "400": {
            "description": "Cookie refresh_token ausente.",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "401": {
            "description": "Refresh token invalido, expirado ou reutilizado.",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        },
        "security": [
          {
            "refreshTokenCookie": []
          }
        ]
      }
    },
    "/users/login": {
      "post": {
        "tags": [
          "Public Project Users"
        ],
        "summary": "Autentica um usuario final via API key",
        "description": "Valida credenciais no projeto associado a API key e define cookies HTTP-only de access token e refresh token.",
        "operationId": "login",
        "parameters": [
          {
            "name": "X-Forwarded-User-Agent",
            "in": "header",
            "description": "User agent real do usuario final repassado pela aplicacao cliente.",
            "required": false,
            "schema": {
              "type": "string"
            },
            "example": "Mozilla/5.0"
          },
          {
            "name": "X-End-User-IP",
            "in": "header",
            "description": "IP real do usuario final repassado pela aplicacao cliente.",
            "required": false,
            "schema": {
              "type": "string"
            },
            "example": "203.0.113.10"
          },
          {
            "name": "X-Device-Name",
            "in": "header",
            "description": "Nome amigavel do dispositivo usado na sessao.",
            "required": false,
            "schema": {
              "type": "string"
            },
            "example": "Chrome macOS"
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UserLoginDTO"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Usuario autenticado. Os cookies access-token e refresh_token sao enviados no header Set-Cookie.",
            "headers": {
              "Set-Cookie": {
                "description": "Define os cookies HTTP-only access-token e refresh_token.",
                "style": "simple",
                "schema": {
                  "type": "string"
                }
              }
            },
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/UserLoginResponseDTO"
                }
              }
            }
          },
          "400": {
            "description": "Corpo da requisicao invalido, erro de validacao ou API key mal formatada.",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "401": {
            "description": "Credenciais invalidas, login bloqueado ou API key invalida.",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "403": {
            "description": "API key sem permissao para autenticar usuarios.",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "409": {
            "description": "Limite de sessoes ativas atingido.",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "500": {
            "description": "Falha ao gerar tokens ou processar API key do projeto.",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        },
        "security": [
          {
            "projectApiKey": []
          }
        ]
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "UpdateUserDTO": {
        "type": "object",
        "properties": {
          "email": {
            "type": "string",
            "format": "email",
            "maxLength": 320,
            "minLength": 0
          },
          "password": {
            "type": "string",
            "maxLength": 255,
            "minLength": 0
          },
          "phone": {
            "$ref": "#/components/schemas/PhoneDTO"
          }
        },
        "required": [
          "email"
        ]
      },
      "UserResponseDTO": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string"
          },
          "projectId": {
            "type": "string"
          },
          "email": {
            "type": "string"
          },
          "status": {
            "type": "string",
            "enum": [
              "ACTIVE",
              "DISABLED",
              "BLOCKED"
            ]
          },
          "createdAt": {
            "type": "string",
            "format": "date-time"
          },
          "updatedAt": {
            "type": "string",
            "format": "date-time"
          },
          "lastLoginAt": {
            "type": "string",
            "format": "date-time"
          },
          "lastPasswordChangedAt": {
            "type": "string",
            "format": "date-time"
          },
          "lastFailedLoginAt": {
            "type": "string",
            "format": "date-time"
          },
          "failedLoginAttempts": {
            "type": "integer",
            "format": "int32"
          },
          "lockedUntil": {
            "type": "string",
            "format": "date-time"
          },
          "lastLoginIpAddress": {
            "type": "string"
          },
          "lastLoginUserAgent": {
            "type": "string"
          },
          "phone": {
            "$ref": "#/components/schemas/PhoneDTO"
          }
        }
      },
      "ErrorResponse": {
        "type": "object",
        "properties": {
          "message": {
            "type": "string"
          },
          "status": {
            "type": "integer",
            "format": "int32"
          },
          "error": {
            "type": "string"
          },
          "time": {
            "type": "string",
            "format": "date-time"
          },
          "fields": {
            "type": "object",
            "additionalProperties": {
              "type": "string"
            }
          }
        }
      },
...
      "CreateUserDTO": {
        "type": "object",
        "properties": {
          "email": {
            "type": "string",
            "format": "email",
            "maxLength": 320,
            "minLength": 0
          },
          "password": {
            "type": "string",
            "maxLength": 255,
            "minLength": 0
          },
          "phone": {
            "$ref": "#/components/schemas/PhoneDTO"
          }
        },
        "required": [
          "email",
          "password"
        ]
      },
      "UserLoginDTO": {
        "type": "object",
        "properties": {
          "email": {
            "type": "string",
            "minLength": 1
          },
          "password": {
            "type": "string",
            "minLength": 1
          }
        },
        "required": [
          "email",
          "password"
        ]
      },
...
    },
    "securitySchemes": {
      "bearerAuth": {
        "type": "http",
        "description": "JWT de conta proprietaria enviado no header Authorization: Bearer <token>.",
        "scheme": "bearer",
        "bearerFormat": "JWT"
      },
      "projectApiKey": {
        "type": "apiKey",
        "description": "API key de projeto usada por clientes terceiros. Tambem pode ser enviada como Authorization: Bearer ya_sk_...",
        "name": "X-API-Key",
        "in": "header"
      },
      "accessTokenCookie": {
        "type": "apiKey",
        "description": "Cookie HTTP-only usado pelo fluxo web para autenticacao.",
        "name": "access-token",
        "in": "cookie"
      },
      "refreshTokenCookie": {
        "type": "apiKey",
        "description": "Cookie HTTP-only usado pelo fluxo web para renovar a sessao.",
        "name": "refresh_token",
        "in": "cookie"
      }
    }
  }
}